Minor improvements - Database saves and sends back a double value for further precision. - Remove some redundant buttons at the ending screen.

+++ client/src/main/java/client/scenes/QuestionCtrl.java
@@ -528,7 +528,7 @@ public class QuestionCtrl {
             return indexCell;
         });
         leaderboardUsernames.setCellValueFactory(e -> new SimpleStringProperty(e.getValue().username));
-        leaderboardScores.setCellValueFactory(e -> new SimpleStringProperty(Integer.toString(e.getValue().score)));
+        leaderboardScores.setCellValueFactory(e -> new SimpleStringProperty(Double.toString(e.getValue().score)));
         emotesUsernameColumn.setCellValueFactory(e -> new SimpleStringProperty(e.getValue().username));
         emotesEmoteColumn.setCellValueFactory(e -> {
             ImageView image = e.getValue().image;
